requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.2.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
